Ignore lists of lists in JSON, reenable feature
The way that lists of lists are represented in the generated code is
as an opaque pointer list. The JSON code needs a typed representation
in order to walk and print the value.

Just ignore for now. Lists of lists aren't widely used because people
have to manage the types themselves.