rpc: stop-the-conn locking
This potentially limits throughput but buys implementation simplicity.
I figure this is an okay tradeoff, since the C++ implementation
indirectly makes the same decision.  It guarantees that only one
goroutine can control the tables at a time.

The tests are flaky with a deadlock on connection close... going to fix
tomorrow.