rpc: fix deadlock in promise pipeline
Before this fix, question.PipelineCall would hold the fulfiller's read
lock while trying to make a call on the promise.  If the remote vat
happened to send its return between the time the lock was acquired and
the sendCall was queued, the return wouldn't be able to resolve, which
would then hang the tasks goroutine.