capnp: embargo clients from a Fulfiller
*** Breaking API change: capnp.ImmediateAnswer takes in a Struct instead
of an object.  It was intended to the whole time.  Unfortunately the rpc
package depended on this, so it's temporarily broken.

Fulfiller now has Reject/Fulfill methods instead of Resolve so that we
can definitively know when the promise is resolved.  The clients in the
struct will be embargoed until the queued calls are delivered.