rpc: eliminate messaging cycle in questions
Since pipeline calls would lock and send a message to the connection,
this could (and would) deadlock.  Now all pipeline calls and cancels
have to pass through the connection, which means only the coordinate
goroutine will resolve a question.

As a bit of code cleanup, appCall now has constructor functions to
ensure that achan always has a buffer.