capnp: extract packing to package
The Compressor/Decompressor are not generally useful, so are better
suited to an internal package.

The Compressor had a precondition of having word-aligned input, so I
replaced it with an in-memory solution.  This causes packing to occur on
the whole message instead of individual segments, which can potentially
reduce size.

I've consolidated all of the tests (and removed redundant ones) into a
single table test.