server: add call queuing
Fixes #3 (mostly).

Since an RPC connection's goroutine is responsible for the call order
guarantee, it must block until Call returns an answer.  However, this
leads to #3, where servers that make RPCs on the connection before
Ack'ing will deadlock.  Either the RPC implementation must add queueing
or the server.  Server seemed like a more reasonable place to put the
functionality.

There is still a possibility of deadlock, but it's only after the queue
is full of unack'd calls.