travis: skip installation of capnp tool
Now that the integration tests don't depend on capnp, skipping capnp
installation improves build times and migrates off legacy infra.

It's possible that this may be revisited in the future, but requiring
developers of this library to install the capnp tool and update the
golden data on changes seems reasonable.