Include packages for the capnproto base schemas.
Includes:

* A script to help generate/clean up the packages
* The schemas (modified to generate working code, see the README)
* The generated packages

Per `capnp/README.md`, we're tracking the versions of the schemas on
capnproto master, since stable is too out of date for sandstorm.

This patch also removes the previous generated packages for schema anc
rpc; now that we're actually shipping packages for all of the base
schemas, we should just use those instead.