rpc: remove shutdown deadlock
shutdown was originally written for Conn.Close and included a wait for
goroutines to report done, as is appropriate.  However, when shutdown
was called from one of the manager's goroutines (in case of an abort,
etc.), the goroutines would deadlock.  Now manager.wait is separated out,
and I made Conn.Wait use the manager.wait function.  This has the
side-effect of making Conn.Wait wait for the goroutines to finish, which
is more correct than the old behavior.

Fixes #45