internal/packed: rewrite packed.Reader
The old implementation was complex and did not correctly handle the case
where Read could return an error and non-zero length data.  The new
reader utilizes bufio.Reader to maintain a buffer, and operates on a
word at a time, with a minimal buffer on top to match the io.Reader
interface.

In my benchmarks, this led to a 3x decompression speedup for not
particularly optimized code.

Fixes #49