rpc: defer teardown when locking c.mu
The panic identified in #54 is a race between dying and teardown.
teardown only starts once Conn.workers is done, so every non-dispatch
goroutine checks the state and adds itself to the c.workers pool.

This conceptually addresses the issue, but I'm not totally confident
that this fixes the bug.  It is hard to write a reliable automated repro
case for this issue.  It seems strictly superior to prior iterations of
the code, so I'll run a few experiments.

Updates #54