Add failing test case for invalid stream header padding.
marshalStreamHeader fails to zero the padding bytes after
the last segment size. This is a problem because e.hdrbuf
is reused between calls to Encode. This leaves garbage in
the header in certain circumstances.

This will be fixed in a future commit.

(See the discussion of #70 for more details of the bug).