capnpc-go: fail if requested file has no import annotation
While this doesn't cause issues for the file being generated, it does
cause issues once the file gets imported by other schemas and it's hard
to decipher the error at that point.  This isn't too onerous, as we're
already asking folks to add the Go.package annotation in all cases.

Fixes #41