capnp: rework writePtr/copyPointer algorithm
Now that we have security checks in place, I'm relying on the depth
limit and read limit of the source message to avoid malicious inputs.
This simplifies the code, as the copy now acts as a pretty normal set of
readPtr/writePtr calls.

While I was investigating, I discovered that the existing algorithm was
very wrong for bit lists.  Added a test and fixed the underlying issue.

Also got around to adding the data-only copy optimization back in for
lists.

This removes the rbtree dependency from capnp.

Fixes #80