capnp: enforce copying occurs in subobjects
This case only occurs with struct list members, but the idea is that as
soon as you start copying a struct, you should also be copying the
subobjects, even if it seems like the subobjects don't need to be
copied.  I added a sticky forceCopy parameter that will trigger the
condition.