capnp: add Message() method to pointer types
This was not originally included because it would increase the size of
the Pointer interface.  Now that all of the pointers are concrete types,
we can add in a convenience method.

Removed Interface.Segment, because the only real reason to get the
segment is to get the message.  I consider it an implementation detail
that Ptr stores seg for space, and could conceivably remove that later.