rpc: multiple Sender messages can be allocated before sending
When the return results are placed in the Sender's allocated message,
the message will need a lifetime that extends beyond the next
NewMessage.  Further, there are times where the contents of a sent
message need to be read later (obtaining clients from promised return
values).  Release is now detached from send in the NewMessage API.

Update pipe to keep track of NewMessage leaks (panics on CloseSend).
Helped chase down a few issues.