capnp: prevent Client release from returning an error
In many places it is undesirable for the error from a client's
destruction to be propagated up the call stack
(e.g. ClientPromise.Fulfill).  In the contexts where I expect capability
destruction to fail (i.e. RPC), then the error can be communicated
out-of-band and is much better handled this way.  Similarly, it forces
locally implemented capabilities to make the same decision to keep an
out-of-band error reporting mechanism.

Since this would be a different contract from io.Closer, the method
names have changed.  Shutdown is for finalizing a ClientHook, and
Release decrements the reference of a Client (possibly calling
Shutdown).