capnp: fix double-far pointer logic
Before this change, both reading and writing double-far pointers were
broken.  Writing a double-far pointer would place a non-zero offset in
the tag word of the landing pad, which the C++ implementation will read
without error.  Reading a double-far pointer was behaving entirely
incorrectly, and would usually result in an error.

This change moves the responsibility of filling in a pointer offset from
Struct and List into writePtr.  Instead of the offset always being given
a pointer address, this computes the offset relative to a base address.
In the case of a double-far pointer, the base address will be zero.

It's worth noting that I picked up on this problem because MultiSegment
started creating more aggressively small segments.  Opened #99 to track.

Fixes #97