capnp: deprecate Struct.Address and List.Address
Especially for List, where the address is at the beginning of the data,
not the composite literal, the return value is not well-defined and its
not clear how to use it.

To replace the usage inside tests, I added a new function, SamePtr.
Tests that need to inspect implementation use implementation details.

I could see adding a method that returns an object's content as a byte
slice, but I will wait for demand arises.