rpc: fix race between importClient.Send and shutdown
shutdown could clear the tables while the question was being sent.
While code could be written defensively to anticipate this (it's not a
strict data race), it's much simpler to reason about each interaction
with the Conn as a task to be waited upon, and then prevent new
interactions from starting if the connection is being shut down.

openSends and bgtasks are consolidated into a single WaitGroup: tasks.
The only case where a send can be unreleased with no open interaction is
an unfinished answer.  This is handled specifically in shutdown.