rpc: make Conn responsible for releasing message caps
Previously, a Conn could fill in the CapTable of a Message created by a
Transport and leave it for the Transport to clean up.  Since releasing a
Transport's message requires the sender lock, this can lead to deadlock:
shutting down an import requires sending a Release message.

Now, the Conn must not send a CapTable back to a Transport or use the
CapTable once the Transport encounters the Message again.  The pipe
transport verifies this invariant.  As a side effect, this behavior
allows the same Message to be reused in the pipe transport.  It is
unclear whether this property is useful outside of pipe.

To satisfy this new constraint, Conn steals the CapTable before it sends
the message back to be observed by the Transport.  It can then release
the capabilities at its convenience (usually after unlocking Conn.mu).