rpc: use explict sender lock calls instead of sendSession
sendSession was the wrong abstraction.  It was only used in answer, but
the usage was wrong, since it would not block to acquire the sender
lock, even though it had been released.

Adds a long explanation of the concurrency concerns and guidelines at
the top of rpc.go.  This explanation changes the working definition of
sender lock as part of this broader understanding.  Holding the sender
lock used to imply not holding Conn.mu.  It is now recognized that there
are situations where it is desirable to hold both, so these are now
orthogonal states.  Function doc comments have been updated to reflect
which they need.