rpc: use less context switches to cancel I/O
Ian Lance Taylor pointed out to me that SetReadDeadline and
SetWriteDeadline not only operate on future calls, but currently blocked
ones too.  I somehow glossed over that.  Previously, I was setting
timeouts in a loop, but with this in mind, setting a deadline of
time.Now() acts like a cancel.

This means that every I/O call now uses a goroutine regardless, but the
scheduler can keep them off the run queue until the Context is Done.