capnp: export Client.State() method instead of Client.Brand()
ClientState collects whether or not the client is still resolving, not
just the brand, and makes it more clear that it's a dynamic value.  This
is crucial in the RPC system for determining whether to send a capability
or a promise.

Brand has been given its own wrapper type to avoid brand-asserting
functions like server.IsServer having to take in an interface{}.  This
eliminates the potential mistake of passing a *capnp.Client to IsServer.