Allow a message to be marked as "safe" to prevent read limiter tracking
There are cases (both as a feature and as a bug-fix) where read limiter tracking
needs to be disabled to prevent hitting traversal limits. Particularly, auto-generated
vars not-only hit traversal limits very quickly (if referenced extensively) but
also don't need this check since the attack this check is supposed to protect
against (amplification attacks), is not a possibilty here.