Fix capnpc-go to use the correct length for nodes in the schema var.
As there can only be `len(fnodes)` of nodes in a schema var, use that
for `NewNodes()` instead of the the number of files. This makes the
output deterministic for compiling a .capnp file on its own, or with
others on the same cmdline.

Fixes https://github.com/capnproto/go-capnproto2/issues/145