Update the bundled schema to those included with capnp-c++ 0.9
Along the way, I automated the addition of the necessary $Go.name
annotations via a new "patch" subcommand to gen.sh (and a corresponding
patch), so that these won't need to be applied manually the next time
we update.

Additionally, the script needed a bit of tweaking to deal with the
fact that upstream now ships schema in subdirectories (namely compat/).