Add new internal/mpsc package.
Per discussion on matrix, we will need an unbounded queue implementation
to go forward with the deadlock fixes. Not currently hooked up to
anything, but the plan would be:

- Spawn a new goroutine per connection, which reads from the queue and
  writes messages to the transport until its context is canceled.
- Everywhere we currently call sendMessage(), instead insert something
  into this queue.

Note that the elements of the queue will not just be *capnp.Message,
becuase of the way the transport works; they may have to be callbacks or
somesuch.