Add some helpers for acquiring locks
This adds a couple helpers "With" and "Without", which  make it easier
to follow where the boundaries of critical sections (and the holes in
them) are, when the helpers are applicable.

This is especially helpful in that using it consistently means places
where the lock management is *not* just a simple straight-line critical
section stand out.