Support arbitrary client metadata, avoid linear seach in sendCap. (#208)
* Support arbitrary client metadata, avoid linear seach in sendCap.

This solves an issue I discovered where sendCap() does a linear search
on the exports table for each sent capability.

After a long discussion with Louis over matrix, we concluded that adding
a metadata object for clients was the best solution; there's a general
design awkwardness right now where the rpc system needs to be able to
inspect the client in various ways, but Client is defined in the main
package and tries to encapsulate its implementation details. To some
degree we are able to use Brand(), but this gets awkward quickly,
because:

1. ClientHook is an openly implementable interface, so we don't know
   all the posibilities in the rpc code.
2. As I've been working on supporting receiverAnswer, I've found
   pinning down all the implementations I defined in the library that
   need to be worried about is challenging.
3. Some of these implementations aren't even exported, or just return
   nil from Brand().

The Metadata object lets the rpc system manage the bookkeeping
information independently of the internals of a client.

I expect to use this for more in the future, but for now this lets
us avoid the linear search by just storing the export id for each
connection in the client's metadata.

* Add wrappers around metadata access for export ids.

This improves type safety at call sites. It's actually more code right
now, but I am working on a branch where I use these elsewhere.

* Drop export ids from metadata when shutting down a connection.

* Add a TODO re: tentative API.

* Fix build error introduced by merge

* Fold Metadata.init() into NewMetadata()

Per discussion with Louis. The wording of Metadata's doc comment has
been updated to also suggest that it is ok to copy Metadata before use.

* Use syncutil.With on metadata