Handle incoming `receiverAnswer` cap descriptors. (#209)
* Cleanup: Stop passing around a superfluous []ClientState

This only actually gets used in sendCap(), but there we have a reference
to the corresponding client, so we can just call client.State() instead
of carrying around this parallel slice everywhere.

* Add a Promise.Resolve helper method.

Useful to avoid needing to expand the if/else in-line everywhere.

* Attach a capnp.Promise to answer table entries.

...in preparation for handling incoming receiverAnswer caps.

* WIP: handle receiverAnswer in recvCap.

Needs testing, and I'm not sure what the right value is to return for
local; need to think about it.

* Add some missing format specifiers.

* WIP test using receiverAnswer

* Fill in the rest of the test.

...which fails; need to debug.

* Remember to wrap the args in a struct.

* Get rid of unnecessary server.Policy.

I think I initially put this in to try to work around deadlock issues
with net.Pipe(), but that didn't work anyway, and is unnecessary now
that the test uses socketpair().

* Update comments re: my current understanding of wrt disembargos

* Fix wrong variable name in comments.

There is no field resultsCapTable, only resultCapTable.

* Support arbitrary client metadata, avoid linear seach in sendCap.

This solves an issue I discovered where sendCap() does a linear search
on the exports table for each sent capability.

After a long discussion with Louis over matrix, we concluded that adding
a metadata object for clients was the best solution; there's a general
design awkwardness right now where the rpc system needs to be able to
inspect the client in various ways, but Client is defined in the main
package and tries to encapsulate its implementation details. To some
degree we are able to use Brand(), but this gets awkward quickly,
because:

1. ClientHook is an openly implementable interface, so we don't know
   all the posibilities in the rpc code.
2. As I've been working on supporting receiverAnswer, I've found
   pinning down all the implementations I defined in the library that
   need to be worried about is challenging.
3. Some of these implementations aren't even exported, or just return
   nil from Brand().

The Metadata object lets the rpc system manage the bookkeeping
information independently of the internals of a client.

I expect to use this for more in the future, but for now this lets
us avoid the linear search by just storing the export id for each
connection in the client's metadata.

* Add wrappers around metadata access for export ids.

This improves type safety at call sites. It's actually more code right
now, but I am working on a branch where I use these elsewhere.

* Fix build error as a result of merge.

* Fuss with recvCap.

At least one case where we should return local = true. Need to think
through the logic for everything else.

* recvCap: Return local = true for receiverAnswer

I have convinced myself that this is necessary, though note that this
is not a complete solution; I need to work out how this interacts with
embargo stuff everywhere else.

* Remember to lock metadata.

* Simplify how we determine whether to embargo

Rather than having recvCap deal with this directly, we now have a separate
routine that just checks if a client is local, which we invoke higher up
in the chain.

This means we're not trying to figure this out based on the cap
descriptor, which means (1) fewer cases, and (2), in the case of
a cap descriptor that denotes a promise or answer, we will be able
to recursively call this on the target -- whose original CapDescriptor
may not be known.

This introduces one minor functional change: ErrorClients are now
treated as local. This means we'll set up embargos for them, which
we don't *need* to do, but it shouldn't be a problem; calling methods
on errors is not a case that is terribly sensitive to the extra round
trip. We could try to rework ErrorClient to avoid this, but it doesn't
seem worth the trouble.

* Add another receiverAnswer test.

This one projects off a call, rather than a bootstrap

* Check for finish before ans.promise.Resolve

I'd meant to fill this in and forgot; thanks to Louis for spotting.