Error client: return the error via Brand() (#215)
* errorClient.Brand: return the error.

I've found this useful in some debugging I'm doing on a separate branch.
It will also make it easier to detect error clients elsewhere in the
code.

* Don't embargo ErrorClients

We didn't used to do this, but started doing so for expidiency when
isLocalClient was introduced. It's not really a big deal, but now that
ErrorClients return the error via Brand(), we can detect this fairly
easily and skip the embargo.