Use the embed package to do less in mktemplates.
...with a bit more work we can probably get rid of the mktemplates
command entirely, but this was low hanging fruit.

Note that this also re-generates some of the .capnp.go files. The
differences are I assume because I am using a new version of the capnp
compiler than when these were last generated? It seems nothing changed
except for the exact bytes of the schema; the actual generated *code*
is the same. We really should have CI make sure these are correct somehow.