Fix a deadlock with FlowLimiters.
...by introducing a far less serious bug. See the comments.

This also fixes *another* deadlock, which is far simpler and dumber, and
was in hindsight probably what we were actually seeing when debugging
this: I noticed while making this change we were treating the returned
ReleaseFunc as an *error*, and checking that it was nil before rigging
up the gotResponse() call -- so it would just deterministically not
ever clear used quota.

*Facepalm*

But hey, I found a sublter, more insidious bug while hunting down a
stupid, deterministic one. That's always nice.