Experiment with structured process lifecycle management.
This patch introduces a `proc` package under internal, with an
abstraction I think will be useful in cleaning up the shutdown logic.

It also includes a rework of the send goroutine (in sender.go) that
reflects how I *want* this to work (notably, the logic for drainQueue
runs in the send gorotuine itself), and how the proc package can be used
to help with that. The new sender is dead code; I intend this as a
starting point for discussion.