Fix broken .Ack() implementation.
Because the cancelCurrentCall channel has a fixed buffer size, if there
is an ongoing call, then trying to insert into this will block,
completely defeating .Ack().

This patch gets rid of the convoluted shutdown logic in favor of just
constructing an appropriate context which is tied to both parent
contexts. I'm not sure why I wrote it this way in the first place.