Get rid of the struct wrappers around list types.
...to make this consistent with structs and capabilities.

This introduces a helper script (./internal/gen/gen.go) to generate a
lot of the boilerplate methods, since not embedding means that we don't
just inherit these anymore.

In the future, it may make sense to move more of the boilerplate into
this script; there's a fair chunk of it beyond what this patch added.