Fix a bug where trySend() would not always re-arm the timer.
If there was no message ready to send, trySend() would not re-arm the
timer -- which makes sense in isolation, but could cause deadlocks if
we later *do* need to re-arm the timer. This reworks things so that
the Limiter doesn't store the timer at all, and it doesn't need
resetting. Instead, we just create a new timer at the top of the main
loop each time we need it.