Delete clearCapTable()
Documentation added in #308 makes this redundant; the ReleaseFuncs
returned by the transport already releases any caps in the message,
and in all cases this was called immediately before releasing the
message as a whole. Not doing this at every call site will make things
less brittle.

Note that there is no functional change here, in particular this is not
fixing a double free bug: since clearCapTable sets CapTable to nil, it
doesn't clash with the call to msg.Reset in the ReleaseFunc.