Fix #306
Also, document the fact that ClientPromise.Fulfill steals the reference.

Mixing this up was the root cause of the issue. Specifically both
Promise.ReleaseClients() and Message.Reset() were releasing the same
clients, causing them to be invalid if later used.

This patch just adds a call to .AddRef() in the appropriate spot, so
that the promise has its own references to the clients, as the API
suggests.