Fix a deadlock.
It is illegal to resolve promises while holding connection locks, and I
was seeing TestHandleReturn_regression deadlock every 6th run or so;
changing this fixes it.

Note: since this makes the resolution itself and the manipulation of
the questions table non-atomic, the ordering is critical:

1. First we remove the entry from the table. This means that if the
remote vat sends us messages pertaining to that question id, they will
simply be rejected instead being acted on. This shouldn't happen
normally, but maybe a malicious vat could get us to panic or something.
2. Second, resolve the promise. After this is complete, we can be sure
the client won't be used.
3. Then, and only then, we can return the question id to the pool for
re-use.