Fix #315
This makes sure we reject any outstanding promises when a connection
shuts down. The basic idea is pretty straightforward: loop over the
questions table and reject everything. There is one thing to be careful
of though: we must make sure we never fulfill or reject a promise before
removing it from the questions table, or we could double-resolve it.
This patch also fixes one remaining place where we were violating this
rule.