Fix a regression introduced by releaseQuestions()
This was manifesting as occasional failures of
TestBootstrapReceiverAnswerRpc, as discussed in #318. When this was
introduced, I attempted to avoid double-rejecting the promise by
removing it from the table after the first rejection, but this is in
fact incorrect in the case where we cancel the question, because the
entry needs to stick around until the return message comes in.

With this patch, instead, we solve the problem by just having
releaseQuestions() check the `finished` flag before calling Reject.