Fix deadlock due to lock ordering inconsistency.
Previously, it was possible for the resolution of a promise to deadlock
with an outstanding call on that promise, since ClientPromise.Fulfill()
(called by Promise.resolve) waits for outstanding calls to complete, and
if a call on a promise is started while the promise is in the pending
resolution state, that call will wait for resolution to complete.

This patch splits the logic from ClientPromise.Fulfill() into two helper
methods, one of which is run in the existing place, and the other, which
contains the wait & shutdown logic, is invoked later, after the
promise's lock is released.