Clarify a comment
Per #306, the call sites where this is unused struck me as suspicious --
but they are actually fine; sendCap already takes care of incrementing
where necessary.

The return value only needed by sendReturn, so that it can arrange to
release these when it receives a finish message with releaseParamCaps =
true.

This patch clarifies the comment to make it clear that the caller is
not responsible for actually applying the refcounts.