Fix some errors in trueValues.
Most glaringly, we were taking the *maximum* bandwidth to be the
bottleneck, not the minimum. As a secondary issue, we were only adding
in delay for the bottleneck link (whether or not its delay was already
accounting for), not all bandwidth limited links.

With these changes, our TraceTest numbers are only off by a bit under
25% -- is that close enough? But my test program still is getting 20x
less bandwidth than it should...