probeBWState: tweak logic for choosing index.
Specifically: move the undesired state to the end of the array, so we
can just pick a random index before then, rather than having to check if
we go the bad one. This also fixes a probably-inconsequential bug in
that the probability of choosing the index after the low one was
slightly higher than the others; it should be a properly uniform
distribution now.