Use a sync.Pool for MultiSegmentArenas
While working on this, I caught a bug in the transport tests: the test
assume that it is safe to call the ReleaseFunc returned from NewMessage
multiple times, which was incidentally true before but undocmented. This
patch documents this invariant, and also ensures that it holds even when
release() actually does something with the arena.