First steps reducing fmt usage.
Per #364, this is a first step on the road to eliminating usage of fmt
from the library; it removes most direct uses of it from the exc
package. The remaining bits are Failedf and friends. I'd like to just
remove these, but that requires changing every downstream call site, and
I wanted to keep this patch small for today; I can tackle that when I
have a bit more time. I don't think it will be too hard; from a quick
skim WrapError covers 99% of the use cases, and the rest won't be hard
to replace with things from strconv.