Use refcounting for results
There is no functional change here yet; this just factors the recounting
we were already doing into a package, and exposes the ability to
refcount to other parts of the code.

This is a stepping stone towards addressing #349; the idea is that the
method receiver should get a reference to this too to keep it alive for
the right amount of time.