Fix tests for {ListAnd}Serve.
If enough is going on, it's not impossible a timeout like this to
trigger spurriously; I managed to get these to fire after ~7 tries by
running the *whole* test suite with the race detector. Probably merging
the code to collect full stack traces for leaks pushed this over the
edge in that scenario.

Let's just take out the timeouts: this gets rid of the possibility of
spurious failures, and real failures will still hang.