Fix #382
This reworks the locking such that newPipelineCallMessage &
newImportCallMessage expect the caller to already be holding the
connection lock, and then removes the call to syncutil.Without in order
to comply with sendMessage's contract.

There is a big downside here: it means that it is no longer OK for
PlaceArgs to call into the RPC subsystem, as it could cause a deadlock.

We're planning on reworking this interface to get rid of PlaceArgs
anyway (see #64), and this kind of code is not that common and generally
easy to avoid, so in the interest of finishing out one batch of
refactoring before starting the next, I am of the opinion that this is
probably the right thing to do.