answer.sendMsg: change locking contract.
We now expect the caller to be holding the lock. This makes more sense
since the other place we use c.sender is also holding the lock.

This lets us get rid of another Unlock/Lock pair, which was unnecessary.

Future work includes moving c.sender inside c.lk, since we never
actually send when not holding the lock anymore.