Improve type safety wrt locking.
This patch introduces a lockedConn type:

```
type lockedConn Conn
```

...and moves all methods that require the caller to hold c.lk onto it.
It provides methods for running functions while holding the lock.

This makes it much harder to introduce bugs where a method is being
called without holding the lock as required, or vice versa.

The patch also fixes a few call sites that were in fact not doing
locking correctly, which were caught by the type checker.