Fix occasional spurious test failure.
Along the lines of 3fc50651bf0b99184418bc046b922d82b19edddc, just with a
different test: occasionally, with the race detector and other tests
running in parallel, I will get a timeout here. But I think that's just
that there's no reason in principle this can't take more than two
seconds, so under extreme circumstances it may. So this patch removes
the timeout; if there's a real bug the channel receive will just hang
and eventually the test-suite-wide timeout will catch it.