Server.handleCalls: deal with ack/Go correctly when shutting down.
Previously, the logic to check for an ack after returning from
handleCall was missing from the loop that drains the queue after
the server's context is canceled. This means that if when draining the
queue, if a method handle calls .Go() before checking the context,
there will be two goroutines servicing method calls when it returns.

This patch folds the shutdown into the same loop, so that the logic is
the same, and so that the shutdown logic respects acks.

Since the queue is meant to be single-consumer, the exact consequences
of this are a bit hard to reason about, though I have noticed that after
fixing this some of the behavior I had attributed to #423 is less
common.