Fix #375 (I think)
This addresses #375; the now-deleted comment about the time gap being OK
has been incorrect at least since we started handling `receiverAnswer`s
specially; the test that was failing was one where a reference to an
answer's promise filed (and transitively pcall) is handed to a method
implementation.

This patch avoids having a gap where pcall & promise are nil by:

- Changing setPipelineCaller so it can be called before we release the
  lock
- Introducing a stand-in type for the pipeline caller that we can set
  pcall to while the RecvCall is ongoing.

Before fixing the bug it took many thousands or tens of thousands of
test runs to trigger it, so while I have not been able to reproduce the
issue after tens of thousands of runs, that alone is inconclusive -- but
I *think* this fixes the bug. It definitely fixes something.